+++ docs/src/changelog.md
@@ -12,6 +12,12 @@ If your currently installed *Quality-time* version is not the latest version, pl
 
 <!-- The line "## <square-bracket>Unreleased</square-bracket>" is replaced by the release/release.py script with the new release version and release date. -->
 
+## [Unreleased]
+
+### Changed
+
+- Replace the report date picker with a panel that contains the report date picker and the settings for number of dates, time between dates, and date order. Closes [#11616](https://github.com/ICTU/quality-time/issues/11616).
+
 ## v5.35.0 - 2025-07-04
 
 ### Fixed